requirements.txt
google-generativeai
langchain-google-genai
langchain
langchain-community
streamlit
python-dotenv
requests
wikipedia